refactor: 공지사항 글쓰기 모달 리팩터링
