+++ src/app/(admin)/_components/modal/PostNoticeModal.tsx
@@ -4,48 +4,13 @@ import { cn } from "@/utils";
 import React from "react";
 import { createPortal } from "react-dom";
 import { RichTextEditor } from "../editor/RichTextEditor";
+import NoticeModalGroup from "./modalElements/NoticeModalGroup";
+import { noticeFormFields } from "../../_constants/noticeModalFields";
 
 const style = {
   boxStyle: "w-full flex items-center border-b border-gray2",
   labelStyle:
     "w-[100px] h-full px-3 py-2 bg-gray1 font-bold text-[14px] leading-5 text-gray8 flex items-center justify-center",
-};
-
-// 폴더 구조 수정 필요
-interface NoticeModalGroupProps {
-  label: string;
-  type: "writer" | "input";
-}
-
-const NoticeModalGroup = ({ label, type }: NoticeModalGroupProps) => {
-  return (
-    <div
-      className={cn(
-        style.boxStyle,
-        type === "writer" ? "h-[40px]" : "h-[56px]"
-      )}
-    >
-      <p className={style.labelStyle}>{label}</p>
-      <div
-        className={cn(
-          "flex-1 px-4 py-2 bg-white text-[14px] leading-5 text-gray8 border-t border-gray2",
-          type === "writer" ? "h-[40px]" : "h-[56px]"
-        )}
-      >
-        {type === "writer" ? (
-          <p className="flex-1 h-full bg-white text-[14px] leading-5 text-gray8">
-            플레이하이브 관리자
-          </p>
-        ) : (
-          <input
-            type="text"
-            placeholder="내용을 입력해주세요."
-            className="w-full h-[40px] rounded-[5px] border p-3 bg-white border-gray3"
-          />
-        )}
-      </div>
-    </div>
-  );
 };
 
 interface PostNoticeModalProps {
@@ -60,13 +25,6 @@ const PostNoticeModal = ({ show, setShow }: PostNoticeModalProps) => {
     e.preventDefault();
     setShow(false);
   };
-
-  // 폴더 구조 수정 필요
-  const noticeFormFields = [
-    { label: "작성자", type: "writer" },
-    { label: "제목", type: "input" },
-    { label: "첨부링크", type: "input" },
-  ] as const;
 
   return createPortal(
     <div
@@ -87,6 +45,7 @@ const PostNoticeModal = ({ show, setShow }: PostNoticeModalProps) => {
               key={index}
               label={field.label}
               type={field.type}
+              modalStyle={style}
             />
           ))}
           <div className={cn(style.boxStyle, "h-full border-b-0")}>
